Added optional displacement-dependant neutral gear RPM increase.
Now, if displacement information is given, revving at neutral gear
should behave more or less similar to real engines. The RPM increase at
neutral gear was modeled as proportional to the engine torque and, now,
inversely proportional to engine displacement. The "bigger" the engine,
the "heavier" it would be. It is know that this is not always true, but
its a "kludge" approximation.