Moved offset of Course::draw pos. arg to traffic pos. in collision check
This fixes weird camera behavior when entering a curve because the
camera was moving earlier than expected since the player's vehicle
position was ahead of the camera's bottom segment. Since the camera
horizontal position is tied to the player's vehicle horizontal offset,
the camera followed the horizontal shifting of the player's. Reverting
this change would, of course, reintroduce the problem of traffic
collision happening past the player, since the player vehicle sprite is
being draw "ahead" of its actual position (which would be the camera's
bottom segment) (the traffic sprite is, however, being drawn on its
correct position). This is worked around now by adding an equivalent
offset on the traffic vehicle's position only when checking for
collision with the player's. This makes collision appear "right", even
if the traffic is not actually on the same road segment of the player's.

A better way to solve this in the future, would be to introduce a offset
to traffic's position when drawing them, or even better, to actually
solve the player's vehicle sprite offset problem (make player's position
match its sprite's position on the road).