Added possibility to choose between constant or dynamic frame duration.
Now the 'sprite_frame_duration' property can be set as "dynamic", which
triggers the vehicle sprite to have dynamic frame duration, proportional
to the vehicle's speed/wheel speed - a behavior that is actually quite
close to the previous one. However, specifying decimal values will set
the frame duration of each sprite as this value (as it used to be), but
without changing duration according to vehicle speed.

Another property, namely 'sprite_animation_speed_factor', specifies a
speed factor applied to the animation speed, which affects the frame
duration in a similar fashion to what the 'sprite_frame_duration' used
to do. It affects dynamic frame duration only.